Remove some additional debug in some of the deeper code.  Add multi-prompt option for VOC
Remove some console spamming out the parameters and datatypes of the layers of the models.  (Knowing this helped solve a memory issue, but it is no longer needed.

Added a multi-prompt process that doesn't unload and re-load the models when using the Array of prompt.   The array of prompt is a strided string array.  The first string is the prompt, the second string is the negative prompt.  The negative prompt string is required, but can be a zero length string.